Added ggplot2 gradient option to choropleth functions.
Faced challenge where I could not provide a gradient scale with
ggplot2 that also applied to Alaska and Hawaii. While the method
described in the vignette kept the scale the same between the US,
AK, and HI, the color limits specified in the scale did not apply.

So: here is code to add a gradient_scale option to the choropleth
functions, that allows for manually-specified gradients.